requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
yt_dlp==2024.5.27
PyNaCl==1.5.0
